chore(replay): extract shared replay utils into @posthog/replay-utils package
Move platform-agnostic replay preprocessing code from the frontend into
a new pnpm workspace package at common/replay-utils/. This includes:

- Mobile replay transformers (wireframe-to-DOM conversion)
- Snapshot processing (chunking, chrome extension stripping, decompression,
  meta event patching, source keys, throttle capturing)
- parseEncodedSnapshots with injectable SnappyDecompressor for
  platform-agnostic binary format detection and decompression orchestration
- SnapshotStore and segment creation
- Injectable ReplayTelemetry interface replacing direct posthog-js dependency
- Break pre-existing import cycles in mobile transformer code
- Standalone smoke tests with jest + @swc/jest
- Docker build context and .dockerignore updated
- Package exports field with subpath entries for deep imports

The frontend wrapper is reduced to a thin adapter that provides the
browser-specific web worker decompressor and posthog-js telemetry.

All 549 session-recordings tests pass.
