+++ .dockerignore
@@ -5,6 +5,7 @@
 !common/hogvm
 !common/esbuilder
 !common/migration_utils
+!common/replay-utils
 !common/plugin_transpiler/*.*
 !common/plugin_transpiler/src
 !common/tailwind
