+++ common/replay-utils/jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+    transform: {
+        '^.+\\.ts$': ['@swc/jest'],
+    },
+    testEnvironment: 'node',
+    clearMocks: true,
+    testMatch: ['<rootDir>/src/**/*.test.ts'],
+}
