+++ common/replay-utils/src/mobile/mobile.types.ts
