+++ common/replay-utils/src/mobile/transformer/colors.ts
