+++ common/replay-utils/src/mobile/transformer/types.ts
