+++ common/replay-utils/src/snapshot-processing/chrome-extension-stripping.ts
@@ -3,11 +3,8 @@ import { fullSnapshotEvent } from '@posthog/rrweb-types'
 import { EventType } from '@posthog/rrweb-types'
 import { serializedNodeWithId } from '@posthog/rrweb-types'
 
-import { RecordingSnapshot } from '~/types'
+import { RecordingSnapshot } from '../types'
 
-// we have seen some chrome extensions
-// that break playback of session recordings
-// let's try to strip them out
 export const CHROME_EXTENSION_DENY_LIST: Record<string, string> = {
     'dji-sru': 'snap and read',
     mloajfnmjckfjbeeofcdaecbelnblden: 'snap and read',
@@ -174,11 +171,6 @@ export function stripChromeExtensionDataFromNode(
 }
 
 export function stripChromeExtensionData(snapshots: RecordingSnapshot[]): RecordingSnapshot[] {
-    // we're going to iterate the snapshots
-    // if we see a full snapshot, we're going to walk the tree of nodes and child nodes
-    // checking for "chrome-extension" in the attributes
-    // if we find it, we're going to remove it and all of its children
-    // we're going to do this in place and return the modified array
     let strippedChromeExtensionData = false
     const matchedExtensions = new Set<string>()
 
@@ -187,8 +179,6 @@ export function stripChromeExtensionData(snapshots: RecordingSnapshot[]): Record
             continue
         }
         const fullSnapshot = snapshot as RecordingSnapshot & fullSnapshotEvent & eventWithTime
-        // it's slightly yucky that we rely on the identity of matchedExtensions here to gather matches
-        // but way simpler than trying to narrow types and return a value
         if (
             stripChromeExtensionDataFromNode(
                 fullSnapshot.data.node,
@@ -201,7 +191,6 @@ export function stripChromeExtensionData(snapshots: RecordingSnapshot[]): Record
     }
 
     if (strippedChromeExtensionData) {
-        // insert a custom snapshot event to indicate that we've stripped the chrome extension data
         const customSnapshot: RecordingSnapshot = {
             type: EventType.Custom,
             data: {
