+++ common/replay-utils/src/snapshot-processing/decompress.ts
@@ -1,13 +1,18 @@
 import { gunzipSync, strFromU8, strToU8 } from 'fflate'
-import posthog from 'posthog-js'
-import { compressedEventWithTime } from 'posthog-js/lib/src/extensions/replay/external/lazy-loaded-session-recorder'
 
-import { IncrementalSource } from '@posthog/rrweb-types'
-import { EventType } from '@posthog/rrweb-types'
+import { EventType, IncrementalSource } from '@posthog/rrweb-types'
 
+import { ReplayTelemetry } from '../telemetry'
 import { throttleCapture } from './throttle-capturing'
 
-function isCompressedEvent(ev: unknown): ev is compressedEventWithTime {
+interface CompressedEventWithTime {
+    cv: string
+    type: number
+    data: any
+    timestamp: number
+}
+
+function isCompressedEvent(ev: unknown): ev is CompressedEventWithTime {
     return typeof ev === 'object' && ev !== null && 'cv' in ev
 }
 
@@ -18,19 +23,7 @@ function unzip(compressedStr: string | undefined): any {
     return JSON.parse(strFromU8(gunzipSync(strToU8(compressedStr, true))))
 }
 
-/**
- *
- * takes an event that might be from web, might be from mobile,
- * and might be partially compressed,
- * and decompresses it when possible
- *
- * you can't return a union of `KnownType | unknown`
- * so even though this returns `eventWithTime | unknown`
- * it has to be typed as only unknown
- *
- * KLUDGE: we shouldn't need so many type assertions on ev.data but TS is not smart enough to figure it out
- */
-export function decompressEvent(ev: unknown, sessionRecordingId: string): unknown {
+export function decompressEvent(ev: unknown, sessionRecordingId: string, telemetry: ReplayTelemetry): unknown {
     try {
         if (isCompressedEvent(ev)) {
             if (ev.cv === '2024-10') {
@@ -70,20 +63,19 @@ export function decompressEvent(ev: unknown, sessionRecordingId: string): unknow
                 }
             } else {
                 throttleCapture(`${sessionRecordingId}-unknown-compressed-event-version`, () => {
-                    posthog.captureException(new Error('Unknown compressed event version'), {
+                    telemetry.captureException(new Error('Unknown compressed event version'), {
                         feature: 'session-recording-compressed-event-decompression',
                         compressedEvent: ev,
                         compressionVersion: ev.cv,
                     })
                 })
-                // probably unplayable but we don't know how to decompress it
                 return ev
             }
         }
         return ev
     } catch (e) {
         throttleCapture(`${sessionRecordingId}-unknown-compressed-event-version`, () => {
-            posthog.captureException((e as Error) || new Error('Could not decompress event'), {
+            telemetry.captureException((e as Error) || new Error('Could not decompress event'), {
                 feature: 'session-recording-compressed-event-decompression',
                 compressedEvent: ev,
             })
