+++ common/replay-utils/src/snapshot-processing/source-key.ts
@@ -1,7 +1,6 @@
-import { SessionRecordingSnapshotSource, SnapshotSourceType } from '~/types'
+import { SessionRecordingSnapshotSource, SnapshotSourceType } from '../types'
 
 export type SourceKey = `${SnapshotSourceType}-${string}`
 export const keyForSource = (source: SessionRecordingSnapshotSource): SourceKey => {
-    // we should always have a blob_key, but in case we don't for some reason, fall back to source
     return `${source.source}-${source.blob_key || source.source}`
 }
