+++ common/replay-utils/src/telemetry.ts
@@ -1,0 +1,9 @@
+export interface ReplayTelemetry {
+    capture(event: string, properties?: Record<string, unknown>): void
+    captureException(error: Error, properties?: Record<string, unknown>): void
+}
+
+export const noOpTelemetry: ReplayTelemetry = {
+    capture: () => {},
+    captureException: () => {},
+}
