+++ common/replay-utils/src/utils.ts
@@ -1,0 +1,7 @@
+export function isObject(candidate: unknown): candidate is Record<string, unknown> {
+    return typeof candidate === 'object' && candidate !== null && !Array.isArray(candidate)
+}
+
+export function isEmptyObject(candidate: unknown): boolean {
+    return isObject(candidate) && Object.keys(candidate).length === 0
+}
