+++ frontend/jest.config.ts
@@ -124,6 +124,8 @@ const config: Config = {
         '^scenes/(.*)$': '<rootDir>/src/scenes/$1',
         '^products/(.*)$': '<rootDir>/../products/$1',
         '^common/(.*)$': '<rootDir>/../common/$1',
+        '^@posthog/replay-utils$': '<rootDir>/../common/replay-utils/src/index.ts',
+        '^@posthog/replay-utils/(.*)$': '<rootDir>/../common/replay-utils/src/$1',
         '^@posthog/shared-onboarding/(.*)$': '<rootDir>/../docs/onboarding/$1',
         '^@posthog/rrweb/es/rrweb': '@posthog/rrweb/dist/rrweb.min.js',
         d3: '<rootDir>/node_modules/d3/dist/d3.min.js',
