+++ frontend/package.json
@@ -99,6 +99,7 @@
         "@posthog/products-visual-review": "workspace:*",
         "@posthog/products-web-analytics": "workspace:*",
         "@posthog/products-workflows": "workspace:*",
+        "@posthog/replay-utils": "workspace:*",
         "@posthog/rrweb": "0.0.26",
         "@posthog/rrweb-plugin-console-record": "0.0.26",
         "@posthog/rrweb-types": "0.0.26",
