+++ frontend/src/scenes/session-recordings/mobile-replay/parsing.test.ts
@@ -1,8 +1,6 @@
-import {
-    parseEncodedSnapshots,
-    processAllSnapshots,
-} from 'scenes/session-recordings/player/snapshot-processing/process-all-snapshots'
-import { keyForSource } from 'scenes/session-recordings/player/snapshot-processing/source-key'
+import { keyForSource, processAllSnapshots } from '@posthog/replay-utils'
+
+import { parseEncodedSnapshots } from 'scenes/session-recordings/player/snapshot-processing/process-all-snapshots'
 
 import { encodedWebSnapshotData } from './__mocks__/encoded-snapshot-data'
