+++ frontend/src/scenes/session-recordings/player/sessionRecordingDataCoordinatorLogic.ts
@@ -3,6 +3,13 @@ import { actions, beforeUnmount, connect, kea, key, listeners, path, props, redu
 import { subscriptions } from 'kea-subscriptions'
 import posthog from 'posthog-js'
 
+import {
+    getHrefFromSnapshot,
+    keyForSource,
+    processAllSnapshots,
+    SnapshotStore,
+    SourceLoadingState,
+} from '@posthog/replay-utils'
 import { EventType, customEvent, eventWithTime } from '@posthog/rrweb-types'
 
 import { Dayjs, dayjs, now } from 'lib/dayjs'
@@ -23,11 +30,7 @@ import { sessionEventsDataLogic } from './sessionEventsDataLogic'
 import { sessionRecordingCommentsLogic } from './sessionRecordingCommentsLogic'
 import type { sessionRecordingDataCoordinatorLogicType } from './sessionRecordingDataCoordinatorLogicType'
 import { sessionRecordingMetaLogic } from './sessionRecordingMetaLogic'
-import { getHrefFromSnapshot } from './snapshot-processing/patch-meta-event'
-import { processAllSnapshots } from './snapshot-processing/process-all-snapshots'
-import { keyForSource } from './snapshot-processing/source-key'
-import { SnapshotStore } from './snapshot-store/SnapshotStore'
-import { SourceLoadingState } from './snapshot-store/types'
+import { posthogTelemetry } from './snapshot-processing/process-all-snapshots'
 import { snapshotDataLogic } from './snapshotDataLogic'
 import { convertSegmentKinds } from './utils/segment-kind-conversion'
 import { createSegments, mapSnapshotsToWindowId } from './utils/segmenter'
@@ -210,7 +213,8 @@ export const sessionRecordingDataCoordinatorLogic = kea<sessionRecordingDataCoor
                 snapshotsBySource,
                 cache.processingCache,
                 values.viewportForTimestamp,
-                props.sessionRecordingId
+                props.sessionRecordingId,
+                posthogTelemetry
             )
 
             breakpoint()
