+++ frontend/src/scenes/session-recordings/player/snapshot-processing/process-all-snapshots.test.ts
@@ -1,112 +1,12 @@
 import { readFileSync } from 'node:fs'
 import { join } from 'node:path'
 
+import { hasAnyWireframes, keyForSource, processAllSnapshots } from '@posthog/replay-utils'
+
 import { RecordingSnapshot, SessionRecordingSnapshotSource } from '~/types'
 
 import { getDecompressionWorkerManager } from './DecompressionWorkerManager'
-import { hasAnyWireframes, parseEncodedSnapshots, processAllSnapshots } from './process-all-snapshots'
-import { keyForSource } from './source-key'
-
-jest.mock('scenes/session-recordings/mobile-replay', () => ({
-    transformEventToWeb: jest.fn((event: any) => {
-        if (event?.type === 2 && event?.data?.wireframes !== undefined) {
-            const firstWireframe = event.data.wireframes[0]
-            const width = firstWireframe?.width || 400
-            const height = firstWireframe?.height || 800
-            return {
-                ...event,
-                data: {
-                    node: {
-                        type: 0,
-                        childNodes: [
-                            {
-                                type: 1,
-                                name: 'html',
-                                id: 2,
-                            },
-                            {
-                                type: 2,
-                                tagName: 'html',
-                                id: 3,
-                                childNodes: [
-                                    {
-                                        type: 2,
-                                        tagName: 'head',
-                                        id: 4,
-                                        childNodes: [],
-                                    },
-                                    {
-                                        type: 2,
-                                        tagName: 'body',
-                                        id: 5,
-                                        attributes: { 'data-rrweb-id': 5 },
-                                        childNodes: [
-                                            {
-                                                type: 2,
-                                                tagName: 'img',
-                                                id: 100,
-                                                attributes: {
-                                                    'data-rrweb-id': 100,
-                                                    width,
-                                                    height,
-                                                    src: 'data:image/png;base64,test',
-                                                },
-                                                childNodes: [],
-                                            },
-                                        ],
-                                    },
-                                ],
-                            },
-                        ],
-                        id: 1,
-                    },
-                    initialOffset: { top: 0, left: 0 },
-                    href: 'https://example.com',
-                },
-            }
-        }
-
-        if (event?.type === 3 && event?.data?.updates && Array.isArray(event.data.updates)) {
-            const updates = event.data.updates
-            if (updates.some((u: any) => u.wireframe)) {
-                const firstUpdate = updates.find((u: any) => u.wireframe)
-                const wireframe = firstUpdate?.wireframe
-                const width = wireframe?.width || 400
-                const height = wireframe?.height || 800
-
-                return {
-                    ...event,
-                    data: {
-                        source: 0,
-                        adds: [
-                            {
-                                parentId: 5,
-                                nextId: null,
-                                node: {
-                                    type: 2,
-                                    tagName: 'img',
-                                    id: 100,
-                                    attributes: {
-                                        'data-rrweb-id': 100,
-                                        width,
-                                        height,
-                                        src: 'data:image/png;base64,test',
-                                    },
-                                    childNodes: [],
-                                },
-                            },
-                        ],
-                        removes: [],
-                        texts: [],
-                        attributes: [],
-                    },
-                }
-            }
-        }
-
-        return event
-    }),
-}))
+import { parseEncodedSnapshots } from './process-all-snapshots'
 
 // Mock the decompression worker manager
 jest.mock('./DecompressionWorkerManager', () => ({
@@ -550,6 +450,7 @@ describe('process all snapshots', () => {
                             updates: [
                                 {
                                     wireframe: {
+                                        id: 12345,
                                         type: 'screenshot',
                                         base64: 'data:image/webp;base64,test',
                                         width: 400,
@@ -651,6 +552,7 @@ describe('process all snapshots', () => {
                         data: {
                             wireframes: [
                                 {
+                                    id: 12345,
                                     type: 'screenshot',
                                     base64: 'data:image/webp;base64,test',
                                     width: 375,
@@ -686,7 +588,7 @@ describe('process all snapshots', () => {
             expect(firstMeta.data).toEqual({
                 width: 375,
                 height: 667,
-                href: 'https://example.com', // From the mock transformer
+                href: 'unknown',
             })
             expect(firstMeta.windowId).toBe(1)
         })
@@ -703,6 +605,7 @@ describe('process all snapshots', () => {
                         data: {
                             wireframes: [
                                 {
+                                    id: 12345,
                                     type: 'screenshot',
                                     base64: 'data:image/webp;base64,test',
                                     width: 375,
@@ -736,7 +639,7 @@ describe('process all snapshots', () => {
             expect(firstMeta.data).toEqual({
                 width: 375,
                 height: 667,
-                href: 'https://example.com', // From the mock transformer
+                href: 'unknown',
             })
         })
 
@@ -793,8 +696,11 @@ describe('process all snapshots', () => {
                             updates: [
                                 {
                                     wireframe: {
+                                        id: 12345,
                                         type: 'screenshot',
                                         base64: 'data:image/webp;base64,test',
+                                        width: 400,
+                                        height: 800,
                                     },
                                 },
                             ],
@@ -833,7 +739,17 @@ describe('process all snapshots', () => {
                         timestamp: 1000,
                         data: {
                             source: 0,
-                            updates: [{ wireframe: { type: 'screenshot' } }],
+                            updates: [
+                                {
+                                    wireframe: {
+                                        id: 12345,
+                                        type: 'screenshot',
+                                        base64: 'data:image/webp;base64,test',
+                                        width: 400,
+                                        height: 800,
+                                    },
+                                },
+                            ],
                         },
                     },
                     {
@@ -841,7 +757,17 @@ describe('process all snapshots', () => {
                         timestamp: 2000,
                         data: {
                             source: 0,
-                            updates: [{ wireframe: { type: 'screenshot' } }],
+                            updates: [
+                                {
+                                    wireframe: {
+                                        id: 12346,
+                                        type: 'screenshot',
+                                        base64: 'data:image/webp;base64,test',
+                                        width: 400,
+                                        height: 800,
+                                    },
+                                },
+                            ],
                         },
                     },
                 ],
@@ -877,6 +803,7 @@ describe('process all snapshots', () => {
                 updates: [
                     {
                         wireframe: {
+                            id: 12345,
                             type: 'screenshot',
                             base64: 'data:image/webp;base64,original-data',
                             width: 400,
@@ -909,7 +836,7 @@ describe('process all snapshots', () => {
                         node: expect.objectContaining({
                             tagName: 'img',
                             attributes: expect.objectContaining({
-                                'data-rrweb-id': expect.any(Number),
+                                'data-rrweb-id': 12345,
                                 width: 400,
                                 height: 800,
                             }),
@@ -933,7 +860,9 @@ describe('process all snapshots', () => {
                             updates: [
                                 {
                                     wireframe: {
+                                        id: 12345,
                                         type: 'screenshot',
+                                        base64: 'data:image/webp;base64,test',
                                         width: 414,
                                         height: 896,
                                     },
@@ -1007,8 +936,11 @@ describe('process all snapshots', () => {
                 sessionId
             )
 
+            // Edge case wireframes (empty base64, zero dimensions) produce placeholder divs,
+            // not img nodes, so no synthetic full snapshot is created — but processing doesn't crash
+            expect(result.length).toBeGreaterThan(0)
             const hasFullSnapshot = result.some((r) => r.type === 2)
-            expect(hasFullSnapshot).toBe(true)
+            expect(hasFullSnapshot).toBe(false)
         })
 
         it('handles missing windowId gracefully', async () => {
