+++ frontend/src/scenes/session-recordings/player/snapshot-store/SnapshotStore.test.ts
@@ -1,8 +1,7 @@
+import { SnapshotStore } from '@posthog/replay-utils'
 import { EventType } from '@posthog/rrweb-types'
 
 import { RecordingSnapshot, SessionRecordingSnapshotSource } from '~/types'
-
-import { SnapshotStore } from './SnapshotStore'
 
 function makeSource(
     index: number,
