+++ frontend/src/scenes/session-recordings/player/utils/segment-kind-conversion.ts
@@ -1,6 +1,6 @@
-import { RecordingSegment } from '~/types'
+import { SnapshotStore } from '@posthog/replay-utils'
 
-import { SnapshotStore } from '../snapshot-store/SnapshotStore'
+import { RecordingSegment } from '~/types'
 
 export function convertSegmentKinds(
     segments: RecordingSegment[],
