+++ pnpm-workspace.yaml
@@ -5,6 +5,7 @@ packages:
     - common/plugin_transpiler
     - common/storybook
     - common/mosaic
+    - common/replay-utils
     - common/tailwind
     - frontend
     - playwright
