+++ tsconfig.json
@@ -15,6 +15,8 @@
             "@common/*": ["../common/*"],
             "@posthog/mosaic": ["../common/mosaic/src/index.ts"],
             "@posthog/mosaic/*": ["../common/mosaic/src/*"],
+            "@posthog/replay-utils": ["../common/replay-utils/src/index.ts"],
+            "@posthog/replay-utils/*": ["../common/replay-utils/src/*"],
             // Needed to access the subpath since torph didn't set up types properly
             "torph/react": ["node_modules/torph/dist/react/index"]
         },
@@ -48,6 +50,7 @@
         "products/**/frontend/**/*",
         "products/**/manifest.tsx",
         "common/esbuilder/*",
+        "common/replay-utils/src/**/*",
         "common/storybook/*",
         "docs/onboarding/**/*"
     ],
